new non-`wasm` download code
The old code had syncronous download behavior because `handles` was not a Vec of futures and the `await` was inside of the for loop. It could interleave with other async tasks, but the downloads would occur in series.

The new code uses `tokio::JoinSet` to spawn at most 8 download tasks at a time and insert them as completed to `raw_assets`

Benchmarks should probably be added to measure the download speeds, but those are tricky to write.